refactor: 검색이나 카테고리 선택 관련 컴포넌트를 따로 분리
